Explicitly set permissions on file created
The mode argument to open() specifies the maximum allowed permissions. The umask
setting is then applied to further restrict the permissions. Hence, the file
that was being created was not allowed to be written by other users. An explicit
chmod solves that problem.

Also, moving the creation of /tmp/index-bench-<>.txt outside of the pipeline
because the pipeline treats this as change in input everytime a new file is
created and keeps re-running the benchmarks. Doing this help it progress to the
slack step :)