Access the Nginx server instead of trying to reach the node server
We don't run the node server in development too, and trying to reach it doesn't
work. This commit changes the development template to change the frontend port
to the nginx server.

Ideally, though, it would be nice to run the node server and access that to
make the development feedback cycle faster.