Run vite server and bbs watcher in local development
This commit adds a script (frontend/scripts/dev.sh) that starts the bsb watcher
and the node server, after making sure that the dependencies are installed.

The VITE_OCAML_BENCH_{PIPELINE,GRAPHQL}_URL variables are now set as
environment variables available at runtime, instead of being available at build
time in the development environment.  The dev.sh script uses these values to
setup the /app/.env file.

For the watcher to detect changes to files correctly, the docker-compose file
has been changed to mount the frontend code to the /app directory, instead of
copying the files as done for the production build.