Add a script to be able to run docker-compose command easily
The project setup uses files in the environment directory, which are
non-standard, and running docker-compose commands requires passing 3 arguments
to get the command to run correctly.  This commit adds a script which makes it
easy to run docker-compose commands.

This would also make it easy to run variations of the start-development and
stop-development commands.

The script can be used to run commands, like

   ./scripts/dc.sh down -v
   ./scripts/dc.sh exec db psql -U docker