Make creating new migrations using pipeline container easier
This commit mounts the ./pipeline/db/migrations directory as a volume in the
pipeline directory. This allows the `omigrate` command to be run in the
pipeline container, making the newly created migration files available in the
host machine immediately.

Previously, the migration files would only be available in the container and
would have to be manually copied to the host machine using the `docker cp`.