Add test_index column to benchmarks table
The field will hold the index of the test in the list of tests in the
benchmarks JSON. This value can be used match the display order of tests with
the order in which they were listed in the generated JSON.

This column has a default value of 0, so that older tests have a value that's
not NULL to avoid problems when trying to order those tests.  There's no way to
reliably add ordering for older tests, unless we write a DB migration to set
the correct test_index value.