Hide benchmarks sidebar menu when there's a single benchmark
The JSON schema supports adding multiple benchmarks, but all of our users
currently have only a single benchmark. The sidebar menu displays this
benchmark's results under the name "Main" which users sometimes confuse to mean
the "main" git branch.

This commit hides the Benchmarks sidebar menu when there's a single benchmark
under which all the tests are present. When there are multiple benchmarks, the
benchmarks MUST to be named and this will make the side bar menu less
confusing.

Closes #136