Pin packages as .dev, unless they specify a version
When a multi-repository project adds a new package with
a constraint `{= "version"}` on an existing one, this will
fail because opam will pin the new package as dev but the
old package as a version of it's last release. Making the
bench-dir folder as a `dev` folder ensures that it's also
picked up as a dev version.