Store metadata in a separate table
This commit is standing on the shoulder of #128. We have introduced
a new table called `benchmarks_metadata` which allows storing only
the metadata. We will eventually phase out of current table, to have
one table with the metadata and another with the metrics. We also
record the build_job_id and run_job_id as and when the build and run
jobs start so we can show the logs in the UI instead of failing as we
do now.