Add make target to clean local test repo
It is useful to clean up the experiments done in local test repo, after
finishing working on a particular feature/bug. It is potentially more
error-prone to do it manually and accidentally deleting the `.git` directory of
the `current-bench` repository. Having a make target for this would make it a
safer action to perform.