Skip old commits
We currently re-run on all open PRs when we re-deploy CB. This means that
sometimes we take more than a day to re-run all benchmarks (a lot of the
older ones will fail because they don't have the changes to run the benchmarks).
This patch skips all commits that are more than two weeks old.