Store PR title in db
Ocurrent fetches all the refs and PRs using the github graphql
interface. The interface is abstracted out so we can get the title
making another graphql query to github. It is not possible to just
make one query because all the types are not exposed by OCurrent,
so we would need to do a lot of copying of types if we wanted to
avoid making another query.