Fix broken frontend watcher screen command
The screen command that runs the `yarn watch` command incorrectly used the `-L`
argument for the log file, instead of the `-Logfile` argument. Not sure how
this has been working so far. Possibly, `node:lts` image changed the base
image, or something which changed the screen version, but I couldn't confirm
this hypothesis.