Convert timeseries data to dygraph format only when plotting
Previously, the timeseries data being handled everywhere in the UI code was
shaped as required by dygraph. This makes the timeseries array much more
complicated than it needs to be. This commit simplifies the shape of the data
being used everywhere and converts it to the dygraph format only when it is
being passed onto the dygraph plotting code.