Add support for multi-value datapoints averaged by the benchmark
Currently, we only support single-value datapoints, or multi-value datapoints
as a list. For a multi-value datapoint, we compute the stats (min, max, and
avg) in the frontend and display the datapoint in the graph. This commit adds
support for multi-value datapoints where these stats have been calculated by
the benchmark itself. The value can be passed as a json object: `{min: number;
max: number; avg: number}`

Closes #132