Change color for delta in metric overview row based on trend
The trend field in the metric metadata is used to indicate whether an increase
in the value is good/bad. When the trend is "higher-is-better" and the value
increases, or the metric is "lower-is-better" and the value decreases, the
delta value is displayed in green. If not, the delta is displayed in red.

Closes #38