Hide benchmarks menu only when an unnamed benchmark is present
Previously, the benchmark menu was incorrectly hidden if there was only one
benchmark for a commit. This commit changes the UI to hide the menu only when
the benchmark is a "default" benchmark which is unnamed. When we have named
benchmarks, we don't hide the menu because the GraphQL requests and ensure that
the GraphQL request has the benchmark name in the request.