frontend: Use same commits on the x-axis for all the metrics
Previously, the x-axis only showed the commits which were present for a
particular metric. This made it hard to read the metric graphs, when a
particular metric fails in some of the runs. This commit changes the graphs to
show all the commits for all the metrics, irrespective of whether a value was
captured for that metric or not for a commit.