frontend: Fix bug with using same x-axis for all metrics
963997b3f0c87147774c04899ac5b0fdd30574a9 added code to ensure that all the
metrics used the same X-axis, even if some metrics didn't have data for
specific commits. The commit incorrectly used the same metadata for all the
commits, and metadata contained other important information like units, trend,
description, etc. This commit fixes the change to use correct metadata for all
the commits for each metric. The metadata from the last commit is used as a
template and the commit and runAt values are changed.