db: Add a postgres view to fetch workers by pull-request
Hasura GraphQL doesn't let us direcly run group by queries. We'd like to group
the benchmarks by workers, docker_images and pull_numbers and fetch the worker
information sorted in the descending order in which the benchmarks were run. To
allow for this we create a new Postgres View - bench_workers. This commit also
adds the corresponding table in the GraphQL metadata and updates the schema json.