pipeline: Don't mark stale PRs as closed PRs
d624b236e3695c399aa6cc5ec366874e87673db2 added functionality to filter out
stale PRs (no activity in 2 weeks) when running the benchmarks on all the
repositories that we are watching. This is to prevent running too many
benchmarks when adding a new repository or when restarting current-bench.

But, e467ef945487ba42c6351cb1da213b5fbf3242dc used the filtered out list of
repositories when trying to figure out the PRs which are open, which is a
bug. This commit fixes that bug.

Closes #361