pipeline: Parse target_version from config in metrics JSON
To support running benchmarks from a different repository than the original
repository, we allow specifying the name and the version of the
original (target project) in the benchmarks config. We capture this version to
a separate column in the benchmarks table in the DB and can be displayed in the
front-end.