pipeline: Allow specifying target version & name as config variables
target_version and target_name are parsed from the JSON output of the
benchmarks run, but this information comes too late.  The DB has unique
constraints on (repo_id, commit, worker and docker_image) on the metadata table
to detect duplicate runs.  This makes it tricky to configure benchmark runs for
a project where the benchmark repository is separate from the code repository.
Including the target_version in the unique constraint above would allow this.