Use literal value in json format example in readme
I imagine there is a tendency to copy this example when making one's
first benchmark, and removing the obvious placeholders (the ellipses).
The problem with including `"value": "benchmark-value"` in the format
example is that it is valid json but isn't accepted by current-bench,
so it may take a while for people to realize that this is the reason
their benchmarks aren't producing graphs.