Support using a separate benchmarks repo for benchmarks (#427)
For repositories like ocaml/ocaml, we would like the benchmarks to be
maintained potentially by a bigger/different set of users, than the ocaml/ocaml
maintainers. This commit adds support for an additional configuration variable
that lets us clone an additional repository and run benchmarks from within it.

NOTE: No new benchmark runs would be triggered by any changes to the benchmarks
repository, with this setup.