Add missing & when invoking the local github-app.sh script
ff71364a6061766e716e4c6deef1e964141682a0 modified the reload.sh script to use
`dune exec --watch` instead of `dune build --watch`. The github-app.sh script
was run in the foreground, with the `dune build` running in the background.
The change to `dune exec` script needs to make the `exec` command to be run in
the foreground, but missed out on making the gitub-app.sh invocation happen in
the background. This commit fixes the problem of `dune exec` not running.