Couple of improvements to using a separate bench_repo (#431)
* Skip installing opam dependencies when using separate bench_repo

* Ensure cached bench_repo is used only when repo has no changes

Simply running git clone as a layer meant that the cached repository is always
used, even when the bench_repo has been changed. To work around this, we use
the GitHub refs API to see if the repo has changed, to decide whether to use
the cached bench_repo or make a new clone of it.  The trick here is from this
GitHub [issue
comment](https://github.com/moby/moby/issues/1996#issuecomment-1152463036)

Also, this commit adds support for specifying a specific branch of the
bench_repo to use for running the benchmarks.  The branch can be specified
using the `/tree/<branch-name>` suffix in the bench_repo URL.