Ensure new builds are triggered on pushes to the `main` branch
With the current app settings, new builds are not triggered when a new commit
is pushed to `main`. Pushes to PR branches do trigger a `POST` to the webhook,
and `current-bench` ends up checking all the branches for updates, which
ensures that the latest commit on `main` is benchmarked too. But, triggers to
`main` don't directly trigger a new build.

Adding read permission on `contents` and enabling webhook triggers for the
`push` event would fix this problem.