Allow for the newly added ngrok domains when upading dev GitHub app
ngrok has added a bunch of new domains earlier this year[1]. Currently, when we
try to use ngrok for local testing, we try to detect the URL of the local app
and update the GitHub app using a POST request. We currently only look for
ngrok.io URLs and fail to find any ngrok URLs with the new domains. This caused
the local development of the app to be broken. This commit makes the search for
ngrok URLs a little more lenient and allows finding the new domains too.

[1]: https://ngrok.com/blog-post/new-ngrok-domains