Prevent conflicts for the dune `_build/.lock` file in the dev env
Currently, we run `dune exec dev/jwt.exe` inside the `github-app.sh` script
which was run as a background process, just before running the build for the
pipeline code in the `watch` mode.  This causes issues with the `jwt.exe`
execution failing because it fails to acquire the `_build/.lock` being held by
the main dune exec process being run in the watch mode.  This commit ensures
that the `jwt.exe` finishes running before we start the dune watcher process.