Include worker and docker_image query params in GitHub status URL
The GitHub status URLs didn't include worker and docker image. When a
repository had multiple configurations or old configurations which were
removed, the default links sometimes may not open the correct page in the app,
which could confuse the users. This commits ensures that the worker and docker
image are included in the URLs to take the users to the correct page from the
PR checks links.

Closes #453